test(delivery): isolate staging retry coverage
